@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <LangVersion>7.3</LangVersion>
     <UserSecretsId>0eefd82a-d34e-4bb9-8bb7-1c0be6bb5380</UserSecretsId>
     <SignAssembly>True</SignAssembly>
